# cmd/ibanator/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for validating IBAN.",
        "title": "IBANator",
        "contact": {}
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/iban/validate": {
            "post": {
                "description": "Runs validation on IBAN. The response gives the validation result, and, if the validation failed, the reason for the failure.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IBAN validation"
                ],
                "summary": "Validates IBAN",
                "parameters": [
                    {
                        "description": "{ ",
                        "name": "IBANValidationRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.IBANValidationRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "iban": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response wrapper",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/main.IBANValidationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Response wrapper containing only an error string",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.IBANValidationRequest": {
            "type": "object",
            "required": [
                "iban"
            ],
            "properties": {
                "iban": {
                    "type": "string"
                }
            }
        },
        "main.IBANValidationResponse": {
            "type": "object",
            "properties": {
                "invalidReason": {
                    "type": "string"
                },
                "valid:": {
                    "type": "boolean"
                }
            }
        },
        "main.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        }
    }
}